+++ cli/.nycrc
@@ -2,5 +2,8 @@
   "all": true,
   "include": [
     "lib/**/*.js"
+  ],
+  "exclude": [
+    "**/*.spec.js"
   ]
 }
