+++ cli/lib/commands/whoami.spec.js
@@ -1,9 +1,9 @@
 const test = require('ava');
-const whoami = require('../../../lib/commands/whoami');
 const sinon = require('sinon');
 
-const FakeApi = require('../../utils/FakeApi');
-const FakeLogger = require('../../utils/FakeLogger');
+const whoami = require('./whoami');
+const FakeApi = require('../../test/utils/FakeApi');
+const FakeLogger = require('../../test/utils/FakeLogger');
 
 test('whoami calls console.log with username when API response is successful', async t => {
   const log = sinon.stub(FakeLogger, 'log');
